Move SummaryListArgs type to govukFrontendTypes
