+++ server/utils/govukFrontendTypes.ts
@@ -28,3 +28,12 @@ export interface InputArgs {
 export interface PanelArgs {
   titleText: string
 }
+
+export interface SummaryListArgs {
+  rows: SummaryListRow[]
+}
+
+interface SummaryListRow {
+  key: { text: string }
+  value: { html?: string; text?: string }
+}
